+++ lib/lightning/policies/users.ex
@@ -1,0 +1,25 @@
+defmodule Lightning.Policies.Users do
+  @moduledoc """
+  The Bodyguard Policy module for users roles.
+  """
+  @behaviour Bodyguard.Policy
+
+  alias Lightning.Accounts.User
+
+  @type actions :: :create_projects
+
+  @doc """
+  authorize/3 takes an action, a user, and a project. It checks the user's role
+  for this project and returns `true` if the user can perform the action and
+  false if they cannot.
+
+  Note that permissions are grouped by action.
+
+  We deny by default, so if a user's role is not added to the allow roles list
+  for a particular action they are denied.
+  """
+  @spec authorize(actions(), Lightning.Accounts.User.t(), any) :: boolean
+  def authorize(:create_projects, %User{role: role}, _project) do
+    role in [:superuser]
+  end
+end